move some logic from App.js to StartScreen.js
